akela-run_test: Check result for Int32

+++ src/akela-llvm/literal.cpp
@@ -23,7 +23,7 @@ namespace Akela_llvm {
                 float v = strtof(n->value.buf, nullptr);
                 return ConstantFP::get(*jd->TheContext, APFloat(v));
             } else if (td->bit_count == 16) {
-                _Float16 v = strtof(n->value.buf, nullptr);
+                _Float16 v = (_Float16)strtof(n->value.buf, nullptr);
                 Type* t = Type::getHalfTy(*jd->TheContext);
                 return ConstantFP::get(t, v);
             } else {
